+++ CHANGELOG.md
@@ -8,6 +8,12 @@ This changelog keeps track of all changes to the Packs SDK. We follow convention
 
 - **Breaking Change** Removed the "autocomplete" property from EmailSchema. It wasn't useful in practice and we want to free up the name "autocomplete" on BaseSchema for better purposes.
 
+## [1.3.4] - 2023-04-17
+
+### Fixed
+
+- Fixed an issue with `placeholder` property being stripped out of the `PropertyIdentifierDetails` schema.
+
 ## [1.3.3] - 2023-04-12
 
 ### Changed
@@ -567,7 +573,7 @@ await myHelper(context);
 
 - Beginning of alpha versioning.
 
-[unreleased]: https://github.com/coda/packs-sdk/compare/v1.3.3...HEAD
+[unreleased]: https://github.com/coda/packs-sdk/compare/v1.3.4...HEAD
 [1.2.3]: https://github.com/coda/packs-sdk/compare/v1.0.1...v1.2.3
 [1.2.2]: https://github.com/coda/packs-sdk/compare/v1.0.1...v1.2.2
 [1.2.1]: https://github.com/coda/packs-sdk/compare/v1.0.1...v1.2.1
@@ -585,3 +591,4 @@ await myHelper(context);
 [0.9.0]: https://github.com/coda/packs-sdk/compare/v0.8.2...v0.9.0
 [1.3.0]: https://github.com/coda/packs-sdk/compare/v1.0.1...v1.3.0
 [1.3.3]: https://github.com/coda/packs-sdk/compare/v1.0.1...v1.3.3
+[1.3.4]: https://github.com/coda/packs-sdk/compare/v1.0.1...v1.3.4